Fix: split reuse & device bind
