+++ component/dialer/bind_darwin.go
@@ -55,19 +55,3 @@ func bindIfaceToListenConfig(ifaceName string, lc *net.ListenConfig, _, address 
 	lc.Control = bindControl(ifaceObj.Index, lc.Control)
 	return address, nil
 }
-
-func addrReuseToListenConfig(lc *net.ListenConfig) {
-	chain := lc.Control
-
-	lc.Control = func(network, address string, c syscall.RawConn) (err error) {
-		defer func() {
-			if err == nil && chain != nil {
-				err = chain(network, address, c)
-			}
-		}()
-
-		return c.Control(func(fd uintptr) {
-			syscall.SetsockoptInt(int(fd), syscall.SOL_SOCKET, syscall.SO_REUSEADDR, 1)
-		})
-	}
-}
