+++ component/dialer/bind_linux.go
@@ -40,19 +40,3 @@ func bindIfaceToListenConfig(ifaceName string, lc *net.ListenConfig, _, address 
 
 	return address, nil
 }
-
-func addrReuseToListenConfig(lc *net.ListenConfig) {
-	chain := lc.Control
-
-	lc.Control = func(network, address string, c syscall.RawConn) (err error) {
-		defer func() {
-			if err == nil && chain != nil {
-				err = chain(network, address, c)
-			}
-		}()
-
-		return c.Control(func(fd uintptr) {
-			syscall.SetsockoptInt(int(fd), syscall.SOL_SOCKET, syscall.SO_REUSEADDR, 1)
-		})
-	}
-}
