+++ component/dialer/bind_others.go
@@ -6,7 +6,6 @@ package dialer
 import (
 	"net"
 	"strconv"
-	"syscall"
 
 	"github.com/Dreamacro/clash/component/iface"
 )
@@ -95,20 +94,4 @@ func bindIfaceToListenConfig(ifaceName string, _ *net.ListenConfig, _, address s
 	}
 
 	return net.JoinHostPort(addr.IP.String(), port), nil
-}
-
-func addrReuseToListenConfig(lc *net.ListenConfig) {
-	chain := lc.Control
-
-	lc.Control = func(network, address string, c syscall.RawConn) (err error) {
-		defer func() {
-			if err == nil && chain != nil {
-				err = chain(network, address, c)
-			}
-		}()
-
-		return c.Control(func(fd uintptr) {
-			syscall.SetsockoptInt(syscall.Handle(fd), syscall.SOL_SOCKET, syscall.SO_REUSEADDR, 1)
-		})
-	}
 }
